+++ tests/unwrap_variant/pass/complex_type.rs
@@ -1,0 +1,11 @@
+use enum_macros::UnwrapVariant;
+
+#[derive(UnwrapVariant)]
+#[allow(dead_code)]
+enum Test<'a, T> {
+    #[unwrap(ref, mut)]
+    A(&'a mut T),
+    B,
+}
+
+fn main() {}